refactor: rename ongoing shipment module
