+++ energetica/init_test_players.py
@@ -2,8 +2,8 @@
 
 from energetica.database.map.hex_tile import HexTile
 from energetica.database.network import Network
+from energetica.database.ongoing_shipment import OngoingShipment
 from energetica.database.player import Player
-from energetica.database.shipment import OngoingShipment
 from energetica.enums import (
     ControllableFacilityType,
     ExtractionFacilityType,
